Refactored styling and repo structure
