+++ src/components/generate-session-component/GenerateSession.jsx
@@ -1,0 +1,21 @@
+import StartSessionButton from '../StartSessionButton';
+import { 
+    Container,
+    Logo,
+    CodeContainer,
+    Text,
+} from './GenerateSession.styles';
+
+function GenerateSession() {
+    return (
+        <Container>
+            <Logo src="./img/callhubLogo-cropped.svg" alt="Callhub Logo" />
+            <CodeContainer>
+                <Text>Employee Side</Text>
+                <StartSessionButton>Generate Session</StartSessionButton>
+            </CodeContainer>
+        </Container>
+    );
+}
+
+export default GenerateSession;
