+++ src/components/generate-session-component/GenerateSession.styles.jsx
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
-import StartSessionButton from './StartSessionButton';
 
-const Container = styled.div`
+export const Container = styled.div`
     background-image: linear-gradient(to bottom right, #0a8e3d, #9fdb3f);
     background-size: contain;
     display: flex;
@@ -11,12 +10,12 @@ const Container = styled.div`
     height: 100vh;
 `;
 
-const Logo = styled.img`
+export const Logo = styled.img`
     height: 270px;
     width: 550px;
 `;
 
-const CodeContainer = styled.div`
+export const CodeContainer = styled.div`
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -29,7 +28,7 @@ const CodeContainer = styled.div`
     box-shadow: 0px 4px 10px #00000040;
 `;
 
-const Text = styled.h3`
+export const Text = styled.h3`
     height: 25px;
     width: 724px;
     color: #000000;
@@ -40,17 +39,3 @@ const Text = styled.h3`
     position: relative;
     margin-bottom: 30px;
 `;
-
-function GenerateSession() {
-    return (
-        <Container>
-            <Logo src="./img/callhubLogo-cropped.svg" alt="Callhub Logo" />
-            <CodeContainer>
-                <Text>Employee Side</Text>
-                <StartSessionButton>Generate Session</StartSessionButton>
-            </CodeContainer>
-        </Container>
-    );
-}
-
-export default GenerateSession;
