+++ src/components/start-session-component/StartSession.jsx
@@ -1,26 +1,10 @@
 import React from 'react';
 import "bootstrap/dist/css/bootstrap.min.css";
 import axios from 'axios';
-import styled from 'styled-components';
-import { connectWebsocket } from '../websocket';
+import { connectWebsocket } from '../../websocket';
 import { useNavigate } from 'react-router-dom';
+import { Button } from './StartSession.styles';
 
-
-const Button = styled.button`
-    height: 50px;
-    width: 60%;
-    background-color: #0b9f43;
-    border-radius: 30px 30px 30px 30px;
-    position: relative;
-    color: white;
-    font-family: 'League Spartan', sans-serif;
-    font-size: x-large;
-    border: 0px solid;
-    cursor: pointer;
-    &:hover {
-        background-color: #076a2d;
-    }
-`
 function StartSessionButton() {
   let navigate = useNavigate();
   const handleSubmit = (e) => {
